pyproject.toml
[project]
name = "llm-fun-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py[cv]>=0.11.0",
    "chromadb>=1.2.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.0.0",
    "ollama>=0.6.0",
    "pgvector>=0.3.6",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.1.1",
    "pyprojroot>=0.3.0",
]
